refactor: メインでコードを置くディレクトリの名前をinternalからappに変更
