+++ cmd/todolist/wire.go
@@ -6,10 +6,10 @@ package main
 import (
 	"github.com/gin-gonic/gin"
 	"github.com/google/wire"
-	"github.com/kzuabe/todolist-go-api/internal/controller"
-	"github.com/kzuabe/todolist-go-api/internal/repository"
-	"github.com/kzuabe/todolist-go-api/internal/router"
-	"github.com/kzuabe/todolist-go-api/internal/usecase"
+	"github.com/kzuabe/todolist-go-api/app/controller"
+	"github.com/kzuabe/todolist-go-api/app/repository"
+	"github.com/kzuabe/todolist-go-api/app/router"
+	"github.com/kzuabe/todolist-go-api/app/usecase"
 	"github.com/kzuabe/todolist-go-api/pkg/middleware"
 )
